ra-tls/elv: Unify LA_REPORT features to one version
One is to remove duplicate code, and the other is to avoid introducing
global variables defined in elv into the LA library.
