+++ ra-tls/elv/echo.go
@@ -1,5 +1,3 @@
-// +build ra
-
 package main
 
 /*
